Added Form and validation to HR view
