+++ package.json
@@ -14,6 +14,7 @@
     "papaparse": "^5.3.2",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-hook-form": "^7.34.0",
     "react-icons": "^4.4.0",
     "react-papaparse": "^4.0.2",
     "react-router-dom": "^6.3.0",
